requirements.txt
flake8==4.0.1
pydantic==1.9.0
psycopg2-binary==2.9.3